# test: Fix `BenchmarkModifySetMutator_Mutate` was broken